Started refactoring libddb to use extern C exports
