+++ src/geoproject.cpp
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "ddb.h"
+#include "dbops.h"
 #include "geoproject.h"
 #include <gdal_priv.h>
 #include <gdal_utils.h>
